Code Review Comments about Ninja Gold assignment
Over all good job on the assignment. 
Take a look over the comments in your code and let me know if you have questions.
Suggest watching your use of session. It is very useful and can be over used when not strictly needed. 
I liked how you came up with changing the colors of your text based on what was being passed in, nice work.